Suggestion box now works! It will create a new file everytime someone makes a suggestion.
For posterity, I added the suggestionbox folder to gitignore so noone else has to read all the terrible things my friends are going to suggest that I do

Anyways yeah I made it do a suggestion box it also shows their username and the time it was posted at. It also doesn't clobber the previous files because it counts how many suggestions there are and recounts it everytime the program starts.