fixed some stuff up in character.py file
i basically made it so that the save function and load function were working nicely together

I made the save and load functions more generic

I made it so the get_profile gives a nice format that similar to what it expects to load in. That way it can convert from human text into computer text and back again and still be readable. Hopefully.